refactor(snapshots-synchronizer): co-locate job runners and trim dead queue API
Merge serial-job-runner.ts and job-queue.ts into a single internal jobs.ts.
They stay two distinct factories (serial lifecycle runner that aborts on stop
vs concurrency-limited thunk queue that drains), now documented as such.

Trim IJobQueue to the only method still used here, scheduleJobWithRetries:
content-file downloads moved to @dcl/content-downloader-component, leaving
scheduleJob / onSizeLessThan / scheduleJobWithPriority / public onIdle dead.
IJobQueue is internal-only, so removing them is safe.
