+++ components/snapshots-synchronizer/src/internal/types.ts
@@ -1,5 +1,5 @@
 import { SnapshotsSynchronizerComponents } from '../types'
-import { IJobQueue } from './job-queue'
+import { IJobQueue } from './jobs'
 
 /**
  * The synchronizer's public dependencies plus the internally-owned concurrency queue used for
